requirements.txt
pytest==3.9.2
python-dotenv==0.16.0
pathlib==1.0.1
Flask==1.1.2
